feat(terminal): allow zooming out to 4 pt
The floor now matches libghostty's own minimum: 4 pt is unreadable, but
it fits a wide TUI on one screen when you only need its shape.

+++ Tests/HerdrMobileTests/TerminalZoomSettingsTests.swift
@@ -52,7 +52,7 @@ struct TerminalZoomSettingsTests {
         #expect(settings.fontSize == 17)
     }
 
-    @Test(arguments: [(Float(-3), Float(8)), (Float(2), Float(8)), (Float(400), Float(32))])
+    @Test(arguments: [(Float(-3), Float(4)), (Float(1), Float(4)), (Float(400), Float(32))])
     func outOfRangeSizesClampToTheSupportedWindow(requested: Float, expected: Float) throws {
         let (defaults, cleanup) = try makeDefaults()
         defer { cleanup() }
